refactor: merge ipa/ipu/tcp-ipu/compatibility tests into one file
test_charm_ipa.py, test_charm_ipu.py, test_charm_ipu_tcp.py,
test_tcp_ipa_compatibility.py, and test_tcp_ipu_compatibility.py each
independently deployed a fresh traefik-k8s (their own spread job/VM/k8s
cluster) just to test one of: ingress-per-app, ingress-per-unit (HTTP
mode), ingress-per-unit (TCP mode), or a pairwise combination of two of
those relating simultaneously.

Merged into tests/integration/test_ingress_relation_types.py: one
shared deployment (traefik + all three testers at once), with the same
16 test functions (renamed only where needed to avoid clashes,
otherwise verbatim) preserving every existing assertion:
- test_ipa_has_ingress / test_ipa_relation_data_shape (from test_charm_ipa.py)
- test_ipu_has_ingress / test_ipu_relation_data_shape (from test_charm_ipu.py)
- test_tcp_relation_data_shape / test_tcp_connection (from test_charm_ipu_tcp.py)
- test_tcp_ipa_compatibility (from test_tcp_ipa_compatibility.py)
- test_tcp_ipu_compatibility (from test_tcp_ipu_compatibility.py)
- test_remove_ipa_relation / test_remove_ipu_relation / test_remove_tcp_relation
  (the individual files' relation-removal checks)

Since all three testers are related sequentially within the same
model before the two compatibility checks run, those checks now also
incidentally prove triple-relation coexistence (IPA+IPU+TCP), not just
the original pairwise scenarios -- a strict superset of prior coverage.

Updated spread.yaml's two explicit MODULE references (the 20.04-base
canary job and the entire arm64 suite, which previously ran only
test_charm_ipa.py) to point at the merged file -- both now exercise
all three ingress types plus their compatibility checks for the same
CI job cost, which is a coverage improvement, not a regression.

Verified via tox -e integration -- --collect-only: 93 tests collected
(down from 103; the 5 merged files' 26 total test functions became 16
in the merged file, accounting for the difference -- no test assertion
was dropped, only the setup/teardown boilerplate was deduplicated).
Full tox -e lint pass clean.
